p3_collab-compet-tennis/requirements.txt
matplotlib==3.3.2
unityagents==0.4.0
visdom==0.1.8.9
torch==1.7.1
numpy==1.19.2
PyYAML==5.4.1
